Add files via upload
An adapter for the T-Rex Arms Sidecar. 
This adapter is intended to be used as a bridge between two attachments to provide a spot for an extra pistol mag. 
I use it between the holster & the tourniquet pouch. However you could use it in many different arrangements. 

For assembly, all that is required is a 1" neodymium magnet. Affix it to the print using JB Weld or other high-strength adhesive. 

https://www.trex-arms.com/product-category/holster-categories/sidecar/

https://www.amazon.com/Realth-Neodymium-Magnets-Permanent-Whiteboard/dp/B0BC9MBW81/ref=sxin_16_pa_sp_search_thematic_sspa?content-id=amzn1.sym.5f50c205-1fbf-463e-8176-d92f3e860d55%3Aamzn1.sym.5f50c205-1fbf-463e-8176-d92f3e860d55&crid=JK01YIXRF3PZ&cv_ct_cx=realth+neodymium+disc+magnets&keywords=realth+neodymium+disc+magnets&pd_rd_i=B0BC9MBW81&pd_rd_r=ff06230b-920d-4c9c-a0e1-55b17d06b66e&pd_rd_w=WyWHw&pd_rd_wg=y4zjI&pf_rd_p=5f50c205-1fbf-463e-8176-d92f3e860d55&pf_rd_r=NTYHAWYMSSD5FHB2PTHQ&qid=1686676349&sprefix=realth+neodymium+disc+magnets%2Caps%2C96&sr=1-1-9cb21890-b538-4fe1-898e-473f8bb6c1fa-spons&sp_csd=d2lkZ2V0TmFtZT1zcF9zZWFyY2hfdGhlbWF0aWM&psc=1